added fd and rd and fixed UI with sidebar

+++ backend/app/schemas/fixed_deposit.py
@@ -1,18 +1,21 @@
-from pydantic import BaseModel, EmailStr
-from decimal import Decimal
+from pydantic import BaseModel
 from datetime import date
-#validates data
+from decimal import Decimal
+
+# Validates FD creation request data
 
 class FDCreate(BaseModel):
     account_id: int
-    amount: float
-    interest_rate: float
+    amount: Decimal
+    interest_rate: Decimal
     start_date: date
     duration_months: int
 
+
+# Defines FD response returned by the API
 class FDResponse(FDCreate):
     fd_id: int
-    maturity_amount: float
+    maturity_amount: Decimal
     status: str
 
     class Config:
